MUCHAS COSAS
-Pago con webpay listo
-Rabbit listo
deberia pdoerse hacer todo lo q tenga q ver con pagos y carrito

+++ server/ms-carrito/src/carrito/carrito.service.ts
@@ -4,9 +4,9 @@ import { Repository } from 'typeorm';
 import { Carrito } from './entities/carrito.entity';
 import { ProductoCarrito } from '../producto-carrito/entities/producto-carrito.entity';
 import { ClientProxy } from '@nestjs/microservices';
-import { Observable } from 'rxjs';
-import { tap } from 'rxjs';
-import { catchError } from 'rxjs';
+import { firstValueFrom } from 'rxjs';
+
+
 
 @Injectable()
 export class CarritoService {
@@ -32,7 +32,24 @@ export class CarritoService {
 
   // Método que agrega un curso al carrito y actualiza el precio
   async addCurso(idCarrito: number, idCurso: number) {
-    
+    // Obtener el precio del curso desde el microservicio de cursos
+    console.log("añadiendo curso al carrito", idCarrito);
+    const cursoPrecio = await firstValueFrom(
+      this.client.send<number>({ cmd: 'get_curso_precio' }, idCurso)
+    ).catch(error => {
+      throw new Error('Error al obtener el precio del curso: ' + error.message);
+    });
+
+    console.log("Precio recibido del curso:", cursoPrecio, "Tipo:", typeof cursoPrecio);
+
+    // Verificar que cursoPrecio es un número
+    const numericCursoPrecio = Number(cursoPrecio);
+    if (isNaN(numericCursoPrecio)) {
+      throw new Error('El precio del curso no es un número válido');
+    }
+
+    console.log("Precio actualizado ibido del curso:", numericCursoPrecio, "Tipo:", typeof numericCursoPrecio);
+
     // Agregar el curso al carrito
     const productoCarrito = new ProductoCarrito();
     productoCarrito.idCarrito = idCarrito;
@@ -43,17 +60,28 @@ export class CarritoService {
     const carrito = await this.carritoRepository.findOneBy({ id: idCarrito });
 
     if (!carrito) {
-      throw new Error('Carrito no encontrado');
+      throw new NotFoundException('Carrito no encontrado');
     }
 
-    // Actualizar el monto 
-    carrito.monto += 100;
+    console.log("Monto del carrito:", carrito.monto, "Tipo:", typeof carrito.monto);
+
+    // Verificar y convertir el monto actual del carrito
+    const numericMonto = Number(carrito.monto);
+    if (isNaN(numericMonto)) {
+      throw new Error('El monto del carrito no es un número válido');
+    }
+    
+
+    // Actualizar el monto
+    carrito.monto = numericCursoPrecio + numericMonto;
     await this.carritoRepository.save(carrito);
 
-    return carrito;  
+    console.log("nuevo precio carrito", carrito.monto);
+
+    return carrito;
   }
 
-  //Eliminar un curso de un carrito
+  // Eliminar un curso de un carrito
   async removeCursoFromCarrito(idCarrito: number, idCurso: number): Promise<void> {
     const curso = await this.productoCarritoRepository.findOne({
       where: { idCarrito, idCurso },
@@ -63,7 +91,43 @@ export class CarritoService {
       throw new NotFoundException('El curso no está en el carrito.');
     }
 
+    // Obtener el precio del curso desde el microservicio de cursos
+    const cursoPrecio = await firstValueFrom(
+      this.client.send<number>({ cmd: 'get_curso_precio' }, idCurso)
+    ).catch(error => {
+      throw new Error('Error al obtener el precio del curso: ' + error.message);
+    });
+
+    console.log("Precio recibido del curso:", cursoPrecio, "Tipo:", typeof cursoPrecio);
+
+    // Verificar que cursoPrecio es un número
+    const numericCursoPrecio = Number(cursoPrecio);
+    if (isNaN(numericCursoPrecio)) {
+      throw new Error('El precio del curso no es un número válido');
+    }
+
     await this.productoCarritoRepository.remove(curso);
+
+    // Obtener el carrito
+    const carrito = await this.carritoRepository.findOneBy({ id: idCarrito });
+
+    if (!carrito) {
+      throw new NotFoundException('Carrito no encontrado');
+    }
+
+    console.log("Monto del carrito:", carrito.monto, "Tipo:", typeof carrito.monto);
+
+    // Verificar y convertir el monto actual del carrito
+    const numericMonto = Number(carrito.monto);
+    if (isNaN(numericMonto)) {
+      throw new Error('El monto del carrito no es un número válido');
+    }
+
+    // Actualizar el monto
+    carrito.monto = numericMonto - numericCursoPrecio;
+    await this.carritoRepository.save(carrito);
+
+    console.log("nuevo precio carrito", carrito.monto);
   }
